separate out paths configurations from main Config class
